requirements.txt
torch>=2.1.0
torchvision>=0.16.0
scikit-image>=0.21.0
scikit-learn>=1.2.0
numpy>=1.23.0
pandas>=1.5.0
matplotlib>=3.7.0
optuna>=3.4.0
Pillow>=9.5.0
kaggle>=1.6.6
